@@ -139,7 +139,7 @@
     "cross-env": "^5.0.0",
     "cross-spawn": "^5.1.0",
     "css-loader": "^0.28.3",
-    "electron": "^1.6.10",
+    "electron": "^1.8.1",
     "electron-builder": "^19.22.1",
     "electron-devtools-installer": "^2.2.0",
     "enzyme": "^2.9.1",
@@ -208,6 +208,7 @@
     "prop-types": "^15.5.10",
     "rc-progress": "^2.1.0",
     "react": "^15.6.1",
+    "react-addons-css-transition-group": "^15.6.2",
     "react-animations": "^0.1.0",
     "react-dom": "^15.6.1",
     "react-grid-layout": "^0.14.4",
